Move tensorizationSettings to MS2DeepscoreSettings
