+++ ms2deepscore/SettingsMS2Deepscore.py
@@ -194,3 +194,28 @@ class GeneratorSettings:
             warnings.warn('When using a fixed set, data will not be shuffled')
         if self.random_seed is not None:
             assert isinstance(self.random_seed, int), "Random seed must be integer number."
+
+
+class TensorizationSettings:
+    """Stores the settings for tensorizing Spectra"""
+    def __init__(self,
+                 **settings):
+        self.min_mz = 10
+        self.max_mz = 1000
+        self.mz_bin_width = 0.1
+        self.intensity_scaling = 0.5
+        self.additional_metadata = []
+        if settings:
+            for key, value in settings.items():
+                if hasattr(self, key):
+                    setattr(self, key, value)
+                else:
+                    raise ValueError(f"Unknown setting: {key}")
+        self.num_bins = int((self.max_mz - self.min_mz) / self.mz_bin_width)
+
+    def get_dict(self):
+        return {"min_mz": self.min_mz,
+                "max_mz": self.max_mz,
+                "mz_bin_width": self.mz_bin_width,
+                "intensity_scaling": self.intensity_scaling,
+                "additional_metadata": self.additional_metadata}
