+++ tests/test_siamese_spectra_model.py
@@ -2,11 +2,10 @@ import numpy as np
 import pytest
 from matchms import Spectrum
 from ms2deepscore.train_new_model.data_generators import (DataGeneratorPytorch,
-                                                          TensorizationSettings,
                                                           tensorize_spectra)
 from ms2deepscore.models.SiameseSpectralModel import (SiameseSpectralModel,
                                                       train)
-from ms2deepscore.SettingsMS2Deepscore import GeneratorSettings
+from ms2deepscore.SettingsMS2Deepscore import GeneratorSettings, TensorizationSettings
 from ms2deepscore.train_new_model.spectrum_pair_selection import \
     select_compound_pairs_wrapper
